// apps/web/src/api/generated/model/transitionTaskDtoTransition.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BTN HRMS API
 * The Human Resource Management System API description
 * OpenAPI spec version: 1.0
 */

export type TransitionTaskDtoTransition = typeof TransitionTaskDtoTransition[keyof typeof TransitionTaskDtoTransition];


export const TransitionTaskDtoTransition = {
  assign: 'assign',
  unassign: 'unassign',
  accept: 'accept',
  reject: 'reject',
  submit: 'submit',
  resubmit: 'resubmit',
  approve: 'approve',
  request_revision: 'request_revision',
  cancel: 'cancel',
} as const;